#16 Moves event handler helpers to a separate trait
